+++ objectbox/lib/src/native/store_config.dart
@@ -1,20 +1,21 @@
 part of 'store.dart';
 
-/// Configuration of a [Store] for use with [WeakStore].
+/// Configuration of a [Store] containing everything required to obtain it
+/// again, e.g. from another isolate.
 class StoreConfiguration {
-  /// [Store._modelDefinition]
-  final ModelDefinition modelDefinition;
-
-  /// [Store.directoryPath]
-  final String directoryPath;
-
-  /// [Store._queriesCaseSensitiveDefault]
-  final bool queriesCaseSensitiveDefault;
-
   /// The ID of the store.
   final int id;
 
+  /// The ModelDefinition of the store.
+  final ModelDefinition modelDefinition;
+
+  /// Path to the database directory.
+  final String directoryPath;
+
+  /// Default value for the string query conditions [caseSensitive] argument.
+  final bool queriesCaseSensitiveDefault;
+
   /// Create a new [StoreConfiguration].
-  StoreConfiguration(this.modelDefinition, this.directoryPath,
-      this.queriesCaseSensitiveDefault, this.id);
+  StoreConfiguration._(this.id, this.modelDefinition, this.directoryPath,
+      this.queriesCaseSensitiveDefault);
 }